# imaxx-smart-office-be-dev/docs/swagger.yaml
basePath: /api
definitions:
...
  enums.ActionRequest:
    enum:
    - save
    - submit
    type: string
    x-enum-varnames:
    - ActionSave
    - ActionSubmit
  enums.LeaveType:
    enum:
    - absence
    - annual
    - sick
    - other
    type: string
    x-enum-varnames:
    - LeaveAbsence
    - LeaveAnnual
    - LeaveSick
    - LeaveOther
...
  request.LeaveFormRequest:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/enums.ActionRequest'
        enum:
        - save
        - submit
      end_date:
        type: string
      leave_type:
        $ref: '#/definitions/enums.LeaveType'
      reason:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - action
    - end_date
    - leave_type
    - reason
    - start_date
    type: object
  request.ReviewRequest:
    properties:
      remark:
        type: string
    type: object
...
  response.DefaultResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
host: localhost:5000
info:
  contact: {}
  description: iMaxx Office API with Swagger
  title: iMaxx Office API
  version: "1.0"
paths:
...
  /leave-forms:
    get:
      consumes:
      - application/json
      description: find all my leave form
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search(state_detail)
        in: query
        name: keyword
        type: string
      - description: Leave Type
        enum:
        - absence
        - annual
        - sick
        in: query
        name: leave_type
        type: string
      - description: Year filter (AD)
        in: query
        name: year
        type: integer
      - description: State (1=Draft, 2=Waiting, 3=Complete, 4=Rejected)
        enum:
        - 1
        - 2
        - 3
        - 4
        in: query
        name: state
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindAll leaveForm
      tags:
      - leaveForms
    post:
      consumes:
      - application/json
      description: create leave form
      parameters:
      - description: create leave form payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LeaveFormRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: CreateLeaveForm leaveform
      tags:
      - leaveForms
  /leave-forms/{id}:
    delete:
      consumes:
      - application/json
      description: delete LeaveForm (only state Draft)
      parameters:
      - description: LeaveForm ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: DeleteLeaveForm leaveForm
      tags:
      - leaveForms
    get:
      consumes:
      - application/json
      description: find by id leaveForm
      parameters:
      - description: LeaveForm ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindById leaveForm
      tags:
      - leaveForms
    put:
      consumes:
      - application/json
      description: update leaveForm save / submit
      parameters:
      - description: update leaveForm payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.LeaveFormRequest'
      - description: LeaveForm ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Update LeaveForm leaveForm
      tags:
      - leaveForms
  /leave-forms/cancel/{id}:
    put:
      consumes:
      - application/json
      description: cancel LeaveForm
      parameters:
      - description: LeaveForm ID (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: cancel leave form and put remark
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: CancelLeaveForm leaveForm
      tags:
      - leaveForms
  /leave-forms/export:
    get:
      consumes:
      - application/json
      description: export to excel
      parameters:
      - description: name search
        in: query
        name: name
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      - description: Leave Type
        enum:
        - absence
        - annual
        - sick
        in: query
        name: leave_type
        type: string
      - description: date filter (start)
        format: date
        in: query
        name: start_date
        type: string
      - description: date filter (end)
        format: date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: Export To Excel leaveform
      tags:
      - leaveForms
  /leave-forms/me:
    get:
      consumes:
      - application/json
      description: find all my leave form
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: keyword search(state_detail)
        in: query
        name: keyword
        type: string
      - description: Leave Type
        enum:
        - absence
        - annual
        - sick
        in: query
        name: leave_type
        type: string
      - description: Year filter (AD)
        in: query
        name: year
        type: integer
      - description: State (1=Draft, 2=Waiting, 3=Complete, 4=Rejected)
        enum:
        - 1
        - 2
        - 3
        - 4
        in: query
        name: state
        type: integer
      - description: leave form uuids
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindMyLeaveAll leaveForm
      tags:
      - leaveForms
  /leave-forms/report:
    get:
      consumes:
      - application/json
      description: find all Report
      parameters:
      - default: 1
        description: page number
        in: query
        name: page
        type: integer
      - default: 10
        description: limit per page number
        in: query
        name: limit
        type: integer
      - default: created_at desc
        description: sort sql string ex. created_at desc
        in: query
        name: sort
        type: string
      - description: name search
        in: query
        name: name
        type: string
      - description: keyword search
        in: query
        name: keyword
        type: string
      - description: Leave Type
        enum:
        - absence
        - annual
        - sick
        in: query
        name: leave_type
        type: string
      - description: date filter (start)
        format: date
        in: query
        name: start_date
        type: string
      - description: date filter (end)
        format: date
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.DefaultResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.DefaultResponse'
      security:
      - BearerAuth: []
      summary: FindReportAll leaveForm
      tags:
      - leaveForms
...
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"